+++ sdk/communication/azure-communication-phonenumbers/azure/communication/phonenumbers/_generated/models/_enums.py
@@ -19,7 +19,7 @@ class BillingFrequency(str, Enum, metaclass=CaseInsensitiveEnumMeta):
 class OperatorNumberType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Type of service associated with the phone number."""
 
-    UNKNOWN = "unknown"
+    UNAVAILABLE = "unavailable"
     OTHER = "other"
     GEOGRAPHIC = "geographic"
     MOBILE = "mobile"
@@ -57,14 +57,33 @@ class PhoneNumberOperationType(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     RELEASE_PHONE_NUMBER = "releasePhoneNumber"
     SEARCH = "search"
     UPDATE_PHONE_NUMBER_CAPABILITIES = "updatePhoneNumberCapabilities"
-    RESERVATION_PURCHASE = "reservationPurchase"
-
 
 class PhoneNumberSource(str, Enum, metaclass=CaseInsensitiveEnumMeta):
     """Source of the number, e.g. Cloud or OperatorConnect."""
 
     CLOUD = "cloud"
     OPERATOR_CONNECT = "operatorConnect"
+
+class PhoneNumberSearchResultError(str, Enum, metaclass=CaseInsensitiveEnumMeta):
+    """Mapping Error Messages to Codes."""
+
+    NO_ERROR = "NoError"
+    UNKNOWN_ERROR_CODE = "UnknownErrorCode"
+    OUT_OF_STOCK = "OutOfStock"
+    AUTHORIZATION_DENIED = "AuthorizationDenied"
+    MISSING_ADDRESS = "MissingAddress"
+    INVALID_ADDRESS = "InvalidAddress"
+    INVALID_OFFER_MODEL = "InvalidOfferModel"
+    NOT_ENOUGH_LICENSES = "NotEnoughLicenses"
+    NO_WALLET = "NoWallet"
+    NOT_ENOUGH_CREDIT = "NotEnoughCredit"
+    NUMBERS_PARTIALLY_ACQUIRED = "NumbersPartiallyAcquired"
+    ALL_NUMBERS_NOT_ACQUIRED = "AllNumbersNotAcquired"
+    RESERVATION_EXPIRED = "ReservationExpired"
+    PURCHASE_FAILED = "PurchaseFailed"
+    BILLING_UNAVAILABLE = "BillingUnavailable"
+    PROVISIONING_FAILED = "ProvisioningFailed"
+    UNKNOWN_SEARCH_ERROR = "UnknownSearchError"
 
 
 class PhoneNumberType(str, Enum, metaclass=CaseInsensitiveEnumMeta):